build: debug symbols: extract upload to s3
